+++ backend/Goalz/Goalz.Domain/Entities/PopUp.cs
@@ -1,0 +1,7 @@
+namespace Goalz.Domain.Entities;
+
+public class PopUp
+{
+    public long Id { get; set; }
+    public string InfoText { get; set; } = string.Empty;
+}